+++ source/blender/editors/include/ED_armature.hh
@@ -301,7 +301,7 @@ bool ED_armature_pose_select_pick_with_buffer(const Scene *scene,
                                               View3D *v3d,
                                               Base *base,
                                               const GPUSelectResult *buffer,
-                                              short hits,
+                                              int hits,
                                               const SelectPick_Params *params,
                                               bool do_nearest);
 /**
